Fix websocket reconnecting issue of kodi component (#26664)
* Fixes the reconnecting issue if, the websocket connection is broken (eg. by power loss)
This is done by pinging the kodi devices in the async_update method. If ping isn't successful,
the component closes the connection and tries to reconnect
* Also fixes the loss of connection if HA was restarted while kodi is running and kodi was after that
rebooted. This is done by enable the poll mode in the component.
If kodi is connected, the polling just pings kodi, and only updates the entity if this is necessary (eg. event play incoming)
* Refactored the connect/reconnect calls